Add access and auth routes for web API * Create Express router objects for access and auth endpoints * Create a middleware setHeaders within access.route.js to set headers,   also use middleware to verify tokens and validate and sanitize input * Use accessController and authController to handle the requests   and send responses * Export the router objects as modules and import them to server.js
This commit adds the functionality of creating access and auth routes
for the web api using Express and middleware. The access route handles
the requests for accessing products and checkout pages (as an example
of public and private routes respectively), while the auth route handles
the requests for signup and login pages.