Reconfigure login/signup response and token
* Added sub, name, email, expires_at, iss and scope to token payload.
* Removed message, user_id, username, email, scope from server response.

This commit added recommended standard properties (as defined by
RFC 7519) to access token. Also removed redundant counterparts because
it increased the amount of data transferred over the network from
response when user sent login or signup request. This made the response
safe as it couldn't reveal the personal data of a customer if the
response was intercepted or compromised. Changed the secret env variable
name for greater consistency.