Add RefreshToken model
* Created refreshToken.model.js file that defines the RefreshToken model
  with user_id, jti and expires_at attributes.
* Added a foreign key constraint on user_id that references the id field
  in the User table.
* Added an association method that sets up a one-to-one relationship
  between RefreshToken and User models.
* Exported the RefreshToken model as a module.

This commit added a new model for storing refresh tokens in the
database. The refresh tokens are used to generate new access tokens when
they expire. The refresh tokens have a unique identifier (jti) and an
expiration date (expires_at). The user_id attribute links the refresh
token to the user who owns it. The association method allows querying
the user data from the refresh token and vice versa. The onDelete option
ensures that the refresh token is removed when the user is deleted from
the database. This makes the authentication system more secure and
efficient.