Update auth.config.js file
* Renamed secret and iss properties to JWT_ACCESS_SECRET and JWT_ISSUER
  respectively.
* Added JWT_REFRESH_SECRET property that holds the secret key for
  signing refresh tokens.
* Added jwtAccessExpiresIn and jwtRefreshExpiresIn properties that
  specify the expiration time for access and refresh tokens in seconds.
* Added userScope property that defines the scope of permissions for
  the user role.

This commit updated the auth.config.js file that contains the
configuration options for the authentication system. The changes made
the file more consistent and descriptive. The new properties allow
customizing the token generation and validation process. The userScope
property can be used to restrict the access of users to certain
resources or actions. This improves the security and flexibility of the
authentication system.