Add expires_at field to db stored refresh token
* Added a line to calculate the expiration date of the refresh token
  using the authConfig.jwtRefreshExpiresIn value

This commit adds the expires_at field to the refresh token to store the
expiration date of the token in the database. This will allow to delete
the expired tokens from the database and prevent unauthorized access.
This commit also fixes the postgresql error that happened when the
expires_at field was not set in the dbToken object.