Add validation schemas for login and signup forms
* Imported Yup library for validation
* Defined username, email and password validation rules
* Exported LoginValidationSchema and SignupValidationSchema objects
* Extended server side validation rules to match client side

This commit adds the validation schemas for login and signup forms using
Yup. It defines the username, email and password validation rules and
exports the LoginValidationSchema and SignupValidationSchema objects for
use in other modules and components. Also it adds additional password
validation rules to express authValidation middleware to match client
validation schema.