Add axios and api for checking user existence
* Installed axios npm package for making HTTP requests
* Added baseURL constant to constants.js file
* Created api.js file with api functions using axios
* Exported checkUserExists function for checking duplicate credentials

This commit adds the axios library. It defines the baseURL constant in
he constants.js file. It also creates the extensible api.js file with
the api functions that use axios to make HTTP requests to the server.
The checkUserExists function is exported for checking if a user with the
same username or email already exists in the database. It also handles
errors and throws them if the response is not successful.