Add async validation for username and email
* Imported checkUserExists function to validationSchemas.js from api.js
* Defined checkDuplicateCredential function using checkUserExists
* Added test method to username and emailSignup schemas with
  checkDuplicateCredential

This commit adds the async validation for username and email in the
signup form using the checkUserExists function from api.js. Whithin
the test method it defines the checkDuplicateCredential function that
takes a value (username or email) and checks if that value is already
existed in the database using checkUserExists. Also this commit adds
validateOnChange={false} prop on Formik component in Login and Signup
that prevents api requests on every value change. This may be changed in
future. For now requests are sent only on blur event.