Add emailSanitize function to authValidation.js
* Defined and exported emailSanitize function to transform email in
  request body.
* Changed check route from GET to POST and added emailSanitize
  middleware.

This commit adds the emailSanitize function to the authValidation
middleware. The purpose of this function is to normalize the email
before checking for duplication in the database while the user fills out
the registration form in the client application.
Additionally, this commit changes the check route from GET to POST
and adds the normalize middleware before the checkController in the
check.route.js file.