Add signed and path options to httpOnly cookies
* Added a secret to sign cookies with cookie parser.
* Cookies are created in two instances with the same value, but for
  different paths.
* In case of errors in refresh token verification cookeis are deleted
  from different paths.

This commit makes the httpOnly cookies for passing the refresh token
signed to make sure they haven't been modified. Due to the fact that
this cookie is needed on different paths, it had to be split in two.
It is likely that the paths will be optimized, in which case there will
be no need for two cookies with the same value.