Move refresh token update logic to auth controller
- Change the route for updating user access from "/api/verify/refresh"
  to "/api/auth/refresh".
- Move the controller responsible for updating access from
  access.controller.refreshAccess to auth.controller.refresh.
- Reduce the amount of code by creating the functions sendAuthResponse
  and getTokensExpirationDates.
- The token.service now exports a function that generates both access
  and refresh tokens - generateTokens.

This commit improves the logic for obtaining a new access token (along
with a refresh token) by the client. As a result of changing the route,
there is no longer a need to split the refresh token httpOnly cookie
into two cookies. This commit also removes a significant amount of
duplicate code.