Refactor authJwt.js, use error handling middleware
* Switched to using errorHandler.js middleware instead of sending
  responses in authJwt.
* Added annotations to document the purpose and parameters of the
  middleware functions.
* Checked if the authorization header starts with "Bearer" before
  splitting and extracting the access token.

This commit refactors the authJwt.js middleware to use the error
handling middleware. This way, it delegates the responsibility of
sending responses to the error handlerand keeps the authJwt middleware
focused on verifying the tokens. It also adds annotations to document
the middleware functions and checks for a valid authorization header
format in verifyAccessToken function.