Change the way refresh tokens are stored
* Rename and refactor refresh token model.
* Refactor generateRefreshToken function of token service.
* Rename and refactor middleware for token verification (authJwt.js).
* Fix minor variable names and typos.

This commit changes the way the server stores refresh tokens. Since
refresh tokens in this app are disposable, there is a need to know which
of them are not allowed to use. A better practice to achieve this is to
use a “blacklist” of revoked tokens instead of a “whitelist” of all
issued JWTs (as it was before) that breaks the core principle of using
JWTs.