Add AuthContext.js for JWT authentication
* Installed jwt-decode package.
* Used React context API to create AuthContext.
* Defined loginUser and logoutUser functions to call login and logout API
  methods.
* Created useAuthContext hook for easy access to auth context.

This commit adds a new file AuthContext.js that handles the
authentication logic using JWT. It uses the React ncontext API to create
a global auth state and provide it to the child components. It also
defines functions to login and logout the user using the API and update
the local storage with the access token. It also creates a custom hook
to access the auth context easily from any component.