Refactor TokenError custom error class
Added options object with clearCookie and isRefetchNeeded properties and
use it in tokenVerification middleware.
The isRefetchNeeded parameter is used to let the client know, that it is
needed to redo the access request with a proper Authorization header.
The client could make such repeated request depending on error's status
code, but there are cases with the same code that do not involve a
repeated request