Refactor client's auth flow
* Created secureApi axios instance and redefine -access requests to be
  send by it.
* Added Interceptor.js component that defines axios interceptors of
  secureApi -access requests and responses.
* Added token.js helper file with getAccessToken function. The function
  gets the access token from localStorage and throws errors according to
  different conditions.

Interceptors should be used instead of validatedToken and verifyAccess
(partly) functions of AuthContext.js to add access token to
Authorization header where necessary. This makes the auth flow more
logical and straightforward. In addition, validatedToken function did
not handle a situation in which the access token somehow wasn't
refreshed and an expired token was sent to the server or no
Authorization header was added to request at all,and server returned an
error. This situation is handled by response interceptor.