Refactor Interceptors
* Deleted token.js file.
* Removed access token's check and refresh in client request
  interceptor.

The client-side access token check and refresh before ${role}-access
requests was redundant and unnecessary. The server-side validation is
the only reliable way to verify the access token. Therefore, the
token.js file was removed and the request interceptor was simplified to
just get the access token from local storage and put it in the
authorization header. The response interceptor will handle the server
response and initiate the access token refresh if needed.